Update drbdwatch.sh
Bug Fix: Cannot properly parse which resource is out of data from /proc/drbd if there are multiple DRBD resource. Changing how we pass the resource to the script.  There is no way to correlate the drbd resource number with the configured resource name (that I can see).  User will pass resource_name:drbd_number as the CLI argument, example to be updated in readme.